requirements.txt
streamlit
pandas
pytesseract
Pillow
PyMuPDF
python-Levenshtein
fuzzywuzzy
python-dotenv
pymongo